refactor(Tooltip/Popover): nested element component
