+++ addon/templates/components/bs-popover.hbs
@@ -2,13 +2,13 @@
 
     {{#ember-wormhole to="ember-bootstrap-modal-container" renderInPlace=_renderInPlace}}
 
-        {{#bs-popover-element id=overlayId placement=_placement fade=fade in=in title=title class=class}}
+        {{#bs-popover/element id=overlayId placement=_placement fade=fade in=in title=title class=class}}
             {{#if hasBlock}}
                 {{yield}}
             {{else}}
                 {{title}}
             {{/if}}
-        {{/bs-popover-element}}
+        {{/bs-popover/element}}
 
     {{/ember-wormhole}}
