+++ addon/templates/components/bs-tooltip.hbs
@@ -2,13 +2,13 @@
 
     {{#ember-wormhole to="ember-bootstrap-modal-container" renderInPlace=_renderInPlace}}
 
-        {{#bs-tooltip-element id=overlayId placement=_placement fade=fade in=in class=class}}
+        {{#bs-tooltip/element id=overlayId placement=_placement fade=fade in=in class=class}}
             {{#if hasBlock}}
                 {{yield}}
             {{else}}
                 {{title}}
             {{/if}}
-        {{/bs-tooltip-element}}
+        {{/bs-tooltip/element}}
 
     {{/ember-wormhole}}
