+++ addon/templates/components/bs-tooltip/element.hbs
