--- app/components/bs-popover-element.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap/components/bs-popover-element';
