--- app/components/bs-tooltip-element.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap/components/bs-tooltip-element';
