+++ app/components/bs-tooltip/element.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap/components/bs-tooltip/element';
