+++ tests/integration/components/bs-popover/element-test.js
@@ -1,16 +1,16 @@
 import {moduleForComponent, test} from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('bs-popover-element', 'Integration | Component | bs popover element', {
+moduleForComponent('bs-popover/element', 'Integration | Component | bs-popover/element', {
   integration: true
 });
 
 test('it has correct markup', function(assert) {
   // Template block usage:
   this.render(hbs`
-    {{#bs-popover-element fade=true in=true title="dummy title"}}
+    {{#bs-popover/element fade=true in=true title="dummy title"}}
       template block text
-    {{/bs-popover-element}}
+    {{/bs-popover/element}}
   `);
 
   assert.ok(this.$(':first-child').hasClass('popover'), 'has popover class');
@@ -25,9 +25,9 @@ test('it has correct markup', function(assert) {
 test('it supports different placements', function(assert) {
   let placements = ['top', 'left', 'bottom', 'right'];
   this.render(hbs`
-    {{#bs-popover-element placement=placement title="dummy title"}}
+    {{#bs-popover/element placement=placement title="dummy title"}}
       template block text
-    {{/bs-popover-element}}
+    {{/bs-popover/element}}
   `);
   placements.forEach((placement) => {
     this.set('placement', placement);
@@ -37,9 +37,9 @@ test('it supports different placements', function(assert) {
 
 test('it passes along class attribute', function(assert) {
   this.render(hbs`
-    {{#bs-popover-element placement=placement title="dummy title" class="wide"}}
+    {{#bs-popover/element placement=placement title="dummy title" class="wide"}}
       template block text
-    {{/bs-popover-element}}
+    {{/bs-popover/element}}
   `);
 
   assert.equal(this.$('.popover.wide').length, 1);
