+++ tests/integration/components/bs-tooltip/element-test.js
@@ -1,16 +1,16 @@
 import {moduleForComponent, test} from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('bs-tooltip-element', 'Integration | Component | bs tooltip element', {
+moduleForComponent('bs-tooltip/element', 'Integration | Component | bs-tooltip/element', {
   integration: true
 });
 
 test('it has correct markup', function(assert) {
   // Template block usage:
   this.render(hbs`
-    {{#bs-tooltip-element fade=true in=true}}
+    {{#bs-tooltip/element fade=true in=true}}
       template block text
-    {{/bs-tooltip-element}}
+    {{/bs-tooltip/element}}
   `);
 
   assert.ok(this.$(':first-child').hasClass('tooltip'), 'has tooltip class');
@@ -24,9 +24,9 @@ test('it has correct markup', function(assert) {
 test('it supports different placements', function(assert) {
   let placements = ['top', 'left', 'bottom', 'right'];
   this.render(hbs`
-    {{#bs-tooltip-element placement=placement}}
+    {{#bs-tooltip/element placement=placement}}
       template block text
-    {{/bs-tooltip-element}}
+    {{/bs-tooltip/element}}
   `);
   placements.forEach((placement) => {
     this.set('placement', placement);
